Reorganizing Math Utilities (#584)
