+++ armi/bookkeeping/report/reportingUtils.py
@@ -30,11 +30,12 @@ import numpy
 
 import armi
 from armi import runLog
-from armi import utils
+from armi.utils import getFileSHA1Hash
 from armi.utils import iterables
 from armi.utils import units
 from armi.utils import textProcessors
 from armi.utils import plotting
+from armi.utils.mathematics import findClosest
 from armi import interfaces
 from armi.bookkeeping import report
 from armi.reactor.flags import Flags
@@ -131,7 +132,7 @@ def writeWelcomeHeaders(o, cs):
             shaHash = (
                 "MISSING"
                 if (not fName or not os.path.exists(fName))
-                else utils.getFileSHA1Hash(fName, digits=10)
+                else getFileSHA1Hash(fName, digits=10)
             )
             inputInfo.append((label, fName, shaHash))
 
@@ -649,9 +650,8 @@ def summarizeZones(core, cs):
     peakAssem = highPow[peakIndex]
 
     avgPFrac = sum(pFracList) / len(pFracList)  # true mean power fraction
-    _avgAssemPFrac, avgIndex = utils.findClosest(
-        pFracList, avgPFrac, indx=True
-    )  # the closest-to-average pfrac in the list
+    # the closest-to-average pfrac in the list
+    _avgAssemPFrac, avgIndex = findClosest(pFracList, avgPFrac, indx=True)
     avgAssem = highPow[avgIndex]  # the actual average assembly
 
     # ok, now need counts, and peak and avg. flow and power in high power region.
