+++ armi/operators/settingsValidation.py
@@ -26,6 +26,7 @@ import os
 import armi
 from armi import runLog, settings, utils
 from armi.utils import pathTools
+from armi.utils.mathematics import expandRepeatedFloats
 from armi.reactor import geometry
 from armi.reactor import systemLayoutInput
 from armi.physics import neutronics
@@ -463,7 +464,7 @@ class Inspector:
 
         def _factorsAreValid(factors, maxVal=1.0):
             try:
-                expandedList = utils.expandRepeatedFloats(factors)
+                expandedList = expandRepeatedFloats(factors)
             except (ValueError, IndexError):
                 return False
             return (
@@ -527,8 +528,8 @@ class Inspector:
         def decayCyclesHaveInputThatWillBeIgnored():
             """Check if there is any decay-related input that will be ignored."""
             try:
-                powerFracs = utils.expandRepeatedFloats(self.cs["powerFractions"])
-                availabilities = utils.expandRepeatedFloats(
+                powerFracs = expandRepeatedFloats(self.cs["powerFractions"])
+                availabilities = expandRepeatedFloats(
                     self.cs["availabilityFactors"]
                 ) or ([self.cs["availabilityFactor"]] * self.cs["nCycles"])
             except:  # pylint: disable=bare-except
