+++ armi/utils/__init__.py
@@ -13,84 +13,32 @@
 # limitations under the License.
 
 """Generic ARMI utilities"""
-import os
-import sys
-import time
-import pickle
-import re
-import pkgutil
-import importlib
-import traceback
-import getpass
-import math
-import datetime
-import tempfile
-import shutil
-import threading
-import subprocess
 import collections
-
+import datetime
+import getpass
 import hashlib
-
-import numpy
-import scipy.optimize as sciopt
+import importlib
+import math
+import os
+import pickle
+import pkgutil
+import re
+import shutil
+import subprocess
+import sys
+import tempfile
+import threading
+import time
+import traceback
 
 import armi
 from armi import runLog
 from armi.utils import iterables
 from armi.utils.flags import Flag
+from armi.utils.mathematics import *  # for backwards compatibility
 
 # Read in file 1 MB at a time to reduce memory burden of reading entire file at once
 _HASH_BUFFER_SIZE = 1024 * 1024
-
-# special pattern to deal with FORTRAN-produced scipats without E, like 3.2234-234
-SCIPAT_SPECIAL = re.compile(r"([+-]?\d*\.\d+)[eEdD]?([+-]\d+)")
-
-
-def coverageReportHelper(config, dataPaths):
-    """
-    Small utility function to generate coverage reports.
-
-    This was created to side-step the difficulties in submitting multi-line python
-    commands on-the-fly.
-
-    This combines data paths and then makes html and xml reports for the
-    fully-combined result.
-    """
-    from coverage import Coverage
-    import coverage
-
-    try:
-        cov = Coverage(config_file=config)
-        if dataPaths:
-            # fun fact: if you combine when there's only one file, it gets deleted.
-            cov.combine(data_paths=dataPaths)
-            cov.save()
-        else:
-            cov.load()
-        cov.html_report()
-        cov.xml_report()
-    except PermissionError as e:
-        # Some file systems have some issues with filenames that start with a '.', such as the
-        # .coverage files. If a permissions error is raised, it likely has something to
-        # do with that. We changed the COVERAGE_RESULTS_FILE in cases.py for this reason.
-        runLog.error(
-            f"There was an issue in generating coverage reports due "
-            f"to the following permissions error: {e}"
-        )
-        # disabled until we figure out the problem.
-        # raise
-    except coverage.misc.CoverageException as e:
-        # This is happening when forming the unit test coverage report. This may be
-        # caused by the TestFixture coverage report gobbling up all of the coverage
-        # files before the UnitTests.cov_report task gets a chance to see them. It may
-        # simply be that we dont want a coverage report generated for the TestFixture.
-        # Something to think about. Either way, we do not want to fail the job just
-        # because of this
-        runLog.error(
-            "There was an issue generating coverage reports "
-            "({}):\n{}".format(type(e), e.args)
-        )
 
 
 def getFileSHA1Hash(filePath, digits=40):
@@ -114,80 +62,6 @@ def getFileSHA1Hash(filePath, digits=40):
     return sha1.hexdigest()[:digits]
 
 
-def efmt(a: str) -> str:
-    r"""Converts string exponential number to another string with just 2 digits in the exponent."""
-    # this assumes that none of our numbers will be more than 1e100 or less than 1e-100...
-    if len(a.split("E")) != 2:
-        two = a.split("e")
-    else:
-        two = a.split("E")
-    # print two
-    exp = two[1]  # this is '+002' or '+02' or something
-
-    if len(exp) == 4:  # it has 3 digits of exponent
-        exp = exp[0] + exp[2:]  # gets rid of the hundred's place digit
-
-    return two[0] + "E" + exp
-
-
-def fixThreeDigitExp(strToFloat: str) -> float:
-    """
-    Convert FORTRAN numbers that cannot be converted into floats.
-
-    Notes
-    -----
-    Converts a number line  "9.03231714805651-101" (no e or E) to "9.03231714805651e-101".
-    Some external depletion kernels currently need this fix. From contact with developer:
-    The notation like 1.0-101 is a FORTRAN thing, with history going back to the 60's.
-    They will only put E before an exponent 99 and below.  Fortran will also read these guys
-    just fine, and they are valid floating point numbers.  It would not be a useful effort,
-    in terms of time, trying to get FORTRAN to behave differently.
-    The approach has been to write a routine in the reading code which will interpret these.
-
-    This helps when the scientific number exponent does not fit.
-    """
-    match = SCIPAT_SPECIAL.match(strToFloat)
-    return float("{}E{}".format(*match.groups()))
-
-
-def findClosest(listToSearch, val, indx=False):
-    r"""
-    find closest item in a list.
-
-    Parameters
-    ----------
-    listToSearch : list
-        The list to search through
-
-    val : float
-        The target value that is being searched for in the list
-
-    indx : bool, optional
-        If true, returns minVal and minIndex, otherwise, just the value
-
-    Returns
-    -------
-    minVal : float
-        The item in the listToSearch that is closest to val
-    minI : int
-        The index of the item in listToSearch that is closest to val. Returned if indx=True.
-
-    """
-    d = float("inf")
-    minVal = None
-    minI = None
-    for i, item in enumerate(listToSearch):
-        if abs(item - val) < d:
-            d = abs(item - val)
-            minVal = item
-            minI = i
-    if indx:
-        return minVal, minI
-    else:
-        # backwards compatibility
-        return minVal
-
-
 def copyWithoutBlocking(src, dest):
     """
     Copy a file in a separate thread to avoid blocking while IO completes.
@@ -199,158 +73,6 @@ def copyWithoutBlocking(src, dest):
     t = threading.Thread(target=shutil.copy, args=(src, dest))
     t.start()
     return t
-
-
-def linearInterpolation(x0, y0, x1, y1, targetX=None, targetY=None):
-    r"""
-    does a linear interpolation (or extrapolation) for y=f(x)
-
-    Parameters
-    ----------
-    x0,y0,x1,y1 : float
-        Coordinates of two points to interpolate between
-
-    targetX : float, optional
-        X value to evaluate the line at
-
-    targetY : float, optional
-        Y value we want to find the x value for (inverse interpolation)
-
-    Returns
-    -------
-    interpY : float
-        The value of y(targetX), if targetX is not None
-
-    interpX : float
-        The value of x where y(x) = targetY (if targetY is not None)
-
-    y = m(x-x0) + b
-
-    x = (y-b)/m
-
-    """
-    if x1 == x0:
-        raise ZeroDivisionError("The x-values are identical. Cannot interpolate.")
-
-    m = (y1 - y0) / (x1 - x0)
-    b = -m * x0 + y0
-
-    if targetX is not None:
-        return m * targetX + b
-    else:
-        return (targetY - b) / m
-
-
-def parabolaFromPoints(p1, p2, p3):
-    r"""
-    find the parabola that passes through three points
-
-    We solve a simultaneous equation with three points.
-
-    A = x1**2 x1 1
-        x2**2 x2 1
-        x3**2 x3 1
-
-    b = y1
-        y2
-        y3
-
-    find coefficients Ax=b
-
-    Parameters
-    ----------
-    p1 : tuple
-        first point (x,y) coordinates
-    p2,p3: tuple, second and third points.
-
-    Returns
-    -------
-    a,b,c coefficients of y=ax^2+bx+c
-
-    """
-
-    A = numpy.array(
-        [[p1[0] ** 2, p1[0], 1], [p2[0] ** 2, p2[0], 1], [p3[0] ** 2, p3[0], 1]]
-    )
-
-    b = numpy.array([[p1[1]], [p2[1]], [p3[1]]])
-    try:
-        x = numpy.linalg.solve(A, b)
-    except:
-        print("Error in parabola {} {}".format(A, b))
-        raise
-
-    return float(x[0]), float(x[1]), float(x[2])
-
-
-def parabolicInterpolation(ap, bp, cp, targetY):
-    r"""
-    Given parabola coefficients, this interpolates the time
-    that would give k=targetK.
-
-    keff = at^2+bt+c
-    We want to solve a*t^2+bt+c-targetK = 0.0 for time.
-    if there are real roots, we should probably take the smallest one
-    because the larger one might be at very high burnup.
-    If there are no real roots, just take the point where the deriv ==0, or
-    2at+b=0, so t = -b/2a
-    The slope of the curve is the solution to 2at+b at whatever t has been determined
-
-    Parameters
-    ----------
-    ap, bp,cp : floats
-        coefficients of a parabola y = ap*x^2 + bp*x + cp
-
-    targetK : float
-        The keff to find the cycle length of
-
-    Returns
-    -------
-    realRoots : list of tuples
-        (root, slope)
-        The best guess of the cycle length that will give k=targetK
-        If no positive root was found, this is the maximum of the curve. In that case,
-        it will be a negative number. If there are two positive roots, there will be two entries.
-
-        slope : float
-            The slope of the keff vs. time curve at t=newTime
-
-    """
-    roots = numpy.roots([ap, bp, cp - targetY])
-    realRoots = []
-    for r in roots:
-        if r.imag == 0 and r.real > 0:
-            realRoots.append((r.real, 2.0 * ap * r.real + bp))
-
-    if not realRoots:
-        # no positive real roots. Take maximum and give up for this cyclic.
-        newTime = -bp / (2 * ap)
-        if newTime < 0:
-            raise RuntimeError("No positive roots or maxima.")
-        slope = 2.0 * ap * newTime + bp
-        newTime = (
-            -newTime
-        )  # return a negative newTime to signal that it is not expected to be critical.
-        realRoots = [(newTime, slope)]
-
-    return realRoots
-
-
-def getFloat(val):
-    r"""returns float version of val, or None if it's impossible. Useful for converting
-    user-input into floats when '' might be possible."""
-    try:
-        newVal = float(val)
-        return newVal
-    except:
-        return None
-
-
-def relErr(v1: float, v2: float) -> float:
-    if v1:
-        return (v2 - v1) / v1
-    else:
-        return -1e99
 
 
 def getTimeStepNum(cycleNumber, subcycleNumber, cs):
@@ -433,12 +155,6 @@ def tryPickleOnAllContents(obj, ignore=None, path=None, verbose=False):
                 print(
                     "{0} in {1} cannot be pickled. It is: {2}. ".format(name, obj, ob)
                 )
-                # traceback.print_exc(limit=0,file=sys.stdout)
-
-
-def tryPickleOnAllContents2(*args, **kwargs):
-    # helper
-    print(doTestPickleOnAllContents2(*args, **kwargs))
 
 
 def doTestPickleOnAllContents2(obj, ignore=None, path=None, verbose=False):
@@ -475,7 +191,7 @@ def doTestPickleOnAllContents2(obj, ignore=None, path=None, verbose=False):
 
 class MyPickler(pickle.Pickler):
     r"""
-    The big guns. This will find your pickle errors if all else fails.
+    This will find your pickle errors if all else fails.
 
     Use with tryPickleOnAllContents3.
     """
@@ -499,7 +215,6 @@ def tryPickleOnAllContents3(obj, ignore=None, path=None, verbose=False):
     to make it work like the other testPickle functions and handle errors, you could.
     But usually you just have to find one unpickleable SOB.
     """
-
     with tempfile.TemporaryFile() as output:
         try:
             MyPickler(output).dump(obj)
@@ -508,9 +223,7 @@ def tryPickleOnAllContents3(obj, ignore=None, path=None, verbose=False):
 
 
 def classesInHierarchy(obj, classCounts, visited=None):
-    """
-    Count the number of instances of each class contained in an objects heirarchy.
-    """
+    """Count the number of instances of each class contained in an objects heirarchy."""
     if not isinstance(classCounts, collections.defaultdict):
         raise TypeError(
             "Need to pass in a default dict for classCounts (it's an out param)"
@@ -550,121 +263,6 @@ def slantSplit(val, ratio, nodes, order="low first"):
     return X
 
 
-def newtonsMethod(
-    func, goal, guess, maxIterations=None, cs=None, positiveGuesses=False
-):
-    r"""
-    Solves a Newton's method with the given function, goal value, and first guess.
-
-    Parameters
-    ----------
-    func : function
-        The function that guess will be changed to try to make it return the goal value.
-
-    goal : float
-        The function will be changed until it's return equals this value.
-
-    guess : float
-        The first guess value to do Newton's method on the func.
-
-    maxIterations : int
-        The maximum number of iterations that the Newton's method will be allowed to perform.
-
-
-    Returns
-    -------
-    ans : float
-        The guess that when input to the func returns the goal.
-
-    """
-
-    def goalFunc(guess, func, positiveGuesses):
-        if positiveGuesses is True:
-            guess = abs(guess)
-        funcVal = func(guess)
-        val = abs(goal - funcVal)
-        return val
-
-    if (maxIterations is None) and (cs is not None):
-        maxIterations = cs["maxNewtonsIterations"]
-
-    # try:
-    ans = float(
-        sciopt.newton(
-            goalFunc,
-            guess,
-            args=(func, positiveGuesses),
-            tol=1.0e-3,
-            maxiter=maxIterations,
-        )
-    )
-
-    if positiveGuesses is True:
-        ans = abs(ans)
-
-    return ans
-
-
-def minimizeScalarFunc(
-    func,
-    goal,
-    guess,
-    maxIterations=None,
-    cs=None,
-    positiveGuesses=False,
-    method=None,
-    tol=1.0e-3,
-):
-    r"""
-    Use scipy minimize with the given function, goal value, and first guess.
-
-    Parameters
-    ----------
-    func : function
-        The function that guess will be changed to try to make it return the goal value.
-
-    goal : float
-        The function will be changed until it's return equals this value.
-
-    guess : float
-        The first guess value to do Newton's method on the func.
-
-    maxIterations : int
-        The maximum number of iterations that the Newton's method will be allowed to perform.
-
-
-    Returns
-    -------
-    ans : float
-        The guess that when input to the func returns the goal.
-
-    """
-
-    def goalFunc(guess, func, positiveGuesses):
-        if positiveGuesses is True:
-            guess = abs(guess)
-        funcVal = func(guess)
-        val = abs(goal - funcVal)
-        return val
-
-    if (maxIterations is None) and (cs is not None):
-        maxIterations = cs["maxNewtonsIterations"]
-
-    X = sciopt.minimize(
-        goalFunc,
-        guess,
-        args=(func, positiveGuesses),
-        method=method,
-        tol=tol,
-        options={"maxiter": maxIterations},
-    )
-    ans = float(X["x"])
-    if positiveGuesses is True:
-        ans = abs(ans)
-
-    return ans
-
-
 def runFunctionFromAllModules(funcName, *args, **kwargs):
     r"""
     Runs funcName on all modules of ARMI, if it exists.
@@ -699,34 +297,6 @@ def runFunctionFromAllModules(funcName, *args, **kwargs):
         except:
             # just print traceback but don't throw an error.
             traceback.print_exc()
-
-
-# TODO: move to pathTools (and reference it here for convenience)
-def mkdir(dirname):
-    r"""
-    Keeps trying to make a directory, outputting whatever errors it encounters,
-    until it is successful.
-
-    Parameters
-    ----------
-    dirname : str
-        Path to the directory to create.
-        What you would normally pass to os.mkdir.
-    """
-    numTimesTried = 0
-    while numTimesTried < 1000:
-        try:
-            os.mkdir(dirname)
-            break
-        except FileExistsError:
-            break
-        except Exception as err:
-            numTimesTried += 1
-            # Only ouput err every 10 times.
-            if numTimesTried % 10 == 0:
-                print(err)
-            # Wait 0.5 seconds, try again.
-            time.sleep(0.5)
 
 
 def prependToList(originalList, listToPrepend):
@@ -866,142 +436,6 @@ def createFormattedStrWithDelimiter(
     return formattedString
 
 
-def rotateXY(x, y, degreesCounterclockwise=None, radiansCounterclockwise=None):
-    """
-    Rotates x, y coordinates
-
-    Parameters
-    ----------
-    x, y : array_like
-        coordinates
-
-    degreesCounterclockwise : float
-        Degrees to rotate in the CCW direction
-
-    radiansCounterclockwise : float
-        Radians to rotate in the CCW direction
-
-    Returns
-    -------
-    xr, yr : array_like
-        the rotated coordinates.
-    """
-
-    if radiansCounterclockwise is None:
-        radiansCounterclockwise = degreesCounterclockwise * math.pi / 180.0
-
-    sinT = math.sin(radiansCounterclockwise)
-    cosT = math.cos(radiansCounterclockwise)
-    rotationMatrix = numpy.array([[cosT, -sinT], [sinT, cosT]])
-    xr, yr = rotationMatrix.dot(numpy.vstack((x, y)))
-    if len(xr) > 1:
-        ## Convert to lists because everyone prefers lists for some reason
-        return xr.tolist(), yr.tolist()
-    else:
-        ## Convert to scalar for consistency with old implementation
-        return xr[0], yr[0]
-
-
-def convertToSlice(x, increment=False):
-    """
-    Convert a int, float, list of ints or floats, None, or slice
-    to a slice. Also optionally increments that slice to make it easy to line
-    up lists that don't start with 0.
-
-    Use this with numpy.array (numpy.ndarray) types to easily get selections of it's elements.
-
-    Parameters
-    ----------
-    x : multiple types allowed.
-        int: select one index.
-        list of int: select these index numbers.
-        None: select all indices.
-        slice: select this slice
-
-    Returns
-    -------
-    slice : slice
-        Returns a slice object that can be used in an array
-        like a[x] to select from its members.
-        Also, the slice has its index numbers decremented by 1.
-        It can also return a numpy array, which can be used
-        to slice other numpy arrays in the same way as a slice.
-
-    Examples
-    --------
-    a = numpy.array([10, 11, 12, 13])
-
-    >>> convertToSlice(2)
-    slice(2, 3, None)
-    >>> a[convertToSlice(2)]
-    array([12])
-
-    >>> convertToSlice(2, increment=-1)
-    slice(1, 2, None)
-    >>> a[convertToSlice(2, increment=-1)]
-    array([11])
-
-    >>> a[convertToSlice(None)]
-    array([10, 11, 12, 13])
-
-
-    >>> a[utils.convertToSlice([1, 3])]
-    array([11, 13])
-
-
-    >>> a[utils.convertToSlice([1, 3], increment=-1)]
-    array([10, 12])
-
-    >>> a[utils.convertToSlice(slice(2, 3, None), increment=-1)]
-    array([11])
-
-    """
-    if increment is False:
-        increment = 0
-
-    if not isinstance(increment, int):
-        raise Exception("increment must be False or an integer in utils.convertToSlice")
-
-    if x is None:
-        x = numpy.s_[:]
-
-    if isinstance(x, list):
-        x = numpy.array(x)
-
-    if isinstance(x, (int, numpy.integer)) or isinstance(x, (float, numpy.floating)):
-        x = slice(int(x), int(x) + 1, None)
-
-    # Correct the slice indices to be group instead of index based.
-    # The energy groups are 1..x and the indices are 0..x-1.
-    if isinstance(x, slice):
-        if x.start is not None:
-            jstart = x.start + increment
-        else:
-            jstart = None
-
-        if x.stop is not None:
-            if isinstance(x.stop, list):
-                jstop = [x + increment for x in x.stop]
-            else:
-                jstop = x.stop + increment
-        else:
-            jstop = None
-
-        jstep = x.step
-
-        return numpy.s_[jstart:jstop:jstep]
-
-    elif isinstance(x, numpy.ndarray):
-        return numpy.array([i + increment for i in x])
-
-    else:
-        raise Exception(
-            (
-                "It is not known how to handle x type: " "{0} in utils.convertToSlice"
-            ).format(type(x))
-        )
-
-
 def plotMatrix(
     matrix,
     fName,
@@ -1016,9 +450,7 @@ def plotMatrix(
     cmap=None,
     figsize=None,
 ):
-    """
-    Plots a matrix
-    """
+    """Plots a matrix"""
     import matplotlib
     import matplotlib.pyplot as plt
 
@@ -1026,8 +458,10 @@ def plotMatrix(
         plt.figure(figsize=figsize)  # dpi=300)
     else:
         plt.figure()
+
     if cmap is None:
         cmap = plt.cm.jet  # @UndefinedVariable  #pylint: disable=no-member
+
     cmap.set_bad("w")
     try:
         matrix = matrix.todense()
@@ -1041,9 +475,9 @@ def plotMatrix(
 
     if title is None:
         title = fName
-    plt.imshow(
-        matrix, cmap=cmap, norm=norm, interpolation="nearest"
-    )  # or bicubic or nearest#,vmin=0, vmax=300)
+
+    # or bicubic or nearest#,vmin=0, vmax=300)
+    plt.imshow(matrix, cmap=cmap, norm=norm, interpolation="nearest")
     plt.colorbar()
     plt.title(title)
     plt.xlabel(xlabel)
@@ -1074,90 +508,6 @@ def userName() -> str:
     return re.sub("^[a-zA-Z]-", "", getpass.getuser())
 
 
-def expandRepeatedFloats(repeatedList):
-    """
-    Return an expanded repeat list.
-
-    Notes
-    -----
-    R char is valid for showing the number of repeats in MCNP. For examples the list:
-    [150,  200, '9R']
-    indicates a 150 day cycle followed by 10 200 day cycles.
-    """
-    nonRepeatList = []
-    for val in repeatedList:
-        isRepeat = False
-        if isinstance(val, str):
-            val = val.upper()
-            if val.count("R") > 1:
-                raise ValueError("List had strings that were not repeats")
-            elif "R" in val:
-                val = val.replace("R", "")
-                isRepeat = True
-        if isRepeat:
-            nonRepeatList += [nonRepeatList[-1]] * int(val)
-        else:
-            nonRepeatList.append(float(val))
-    return nonRepeatList
-
-
-def getStepsFromValues(values, prevValue=0.0):
-    """Convert list of floats to list of steps between each float."""
-    steps = []
-    for val in values:
-        currentVal = float(val)
-        steps.append(currentVal - prevValue)
-        prevValue = currentVal
-    return steps
-
-
-def average1DWithinTolerance(vals, tolerance=0.2):
-    """
-    Compute the average of a series of arrays with a tolerance.
-
-    Tuned for averaging assembly meshes or block heights.
-
-    Parameters
-    ----------
-    vals : 2D numpy.array
-        could be assembly x axial mesh tops or heights
-    """
-    vals = numpy.array(vals)
-
-    filterOut = numpy.array([False])  # this gets discarded
-    while not filterOut.all():  # 20% difference is the default tolerance
-        avg = vals.mean(axis=0)  # average over all columns
-        diff = abs(vals - avg) / avg  # no nans, because all vals are non-zero
-        filterOut = (diff > tolerance).sum(
-            axis=1
-        ) == 0  # True = 1, sum across axis means any height in assem is off
-        vals = vals[filterOut]  # filter anything that is skewing
-
-    if vals.size == 0:
-        raise ValueError("Nothing was near the mean, there are no acceptable values!")
-
-    if (avg <= 0.0).any():
-        raise ValueError(
-            "A non-physical value (<=0) was computed, but this is not possible.\n"
-            "Values: {}\navg: {}".format(vals, avg)
-        )
-
-    return avg
-
-
-def findNearestValue(searchList, searchValue):
-    """Search a given list for the value that is closest to the given search value."""
-    return findNearestValueAndIndex(searchList, searchValue)[0]
-
-
-def findNearestValueAndIndex(searchList, searchValue):
-    """Search a given list for the value that is closest to the given search value. Return a tuple
-    containing the value and its index in the list."""
-    searchArray = numpy.array(searchList)
-    closestValueIndex = (numpy.abs(searchArray - searchValue)).argmin()
-    return searchArray[closestValueIndex], closestValueIndex
-
-
 class MergeableDict(dict):
     """
     Overrides python dictionary and implements a merge method.
@@ -1170,9 +520,6 @@ class MergeableDict(dict):
     def merge(self, *otherDictionaries) -> None:
         for dictionary in otherDictionaries:
             self.update(dictionary)
-
-
-shutil_copy = shutil.copy
 
 
 def safeCopy(src: str, dst: str) -> None:
@@ -1191,4 +538,6 @@ def safeCopy(src: str, dst: str) -> None:
     runLog.extra("Copied {} -> {}".format(src, dst))
 
 
+# Allow us to check the copy operation is complete before continuing
+shutil_copy = shutil.copy
 shutil.copy = safeCopy
